fixed topic events

+++ src/events/topics/no-bots.js
@@ -3,6 +3,7 @@ module.exports = {
 }
 
 module.exports.call = async (bot, msg) => {
+    if(msg.channel.type == 'dm') return;
     if(msg.channel.topicSetting ? msg.channel.topicSetting('no-bots') : false){
         if(msg.author.bot) msg.delete().catch(()=>{})
     }
