+++ src/events/topics/one-word.js
@@ -3,6 +3,7 @@ module.exports = {
 }
 
 module.exports.call = async (bot, msg) => {
+    if(msg.channel.type == 'dm') return;
     if(msg.channel.topicSetting('one-word')){
         if(msg.content.match(/\s/g)){
             msg.delete().catch(()=>{})
